final commit

+++ components/HeaderHome.js
@@ -8,11 +8,12 @@ function HeaderHome({ modal }) {
   const router = useRouter()
 
   const logOut = async () => {
-    const user = await axios.get('../api/auth/log')
-    console.log(user)
-    if (user.status === 200) {
-      localStorage.setItem('USERNAME', JSON.stringify(''))
-      router.replace('/login')
+    if (confirm('Are you sure you want to log out ⚠️')) {
+      const user = await axios.get('../api/auth/log')
+      if (user.status === 200) {
+        localStorage.setItem('USERNAME', JSON.stringify(''))
+        router.replace('/login')
+      }
     }
   }
